GUI Prettyness
Making things look nice

+++ src/main/java/Forecast24Hour.java
@@ -86,10 +86,10 @@ public class Forecast24Hour extends TransparentPanel {
 			sclabel[i].setFont(minorfont);
 
 			if (preferences.getTempUnit().equalsIgnoreCase("F")) {
-				templabel[i] = new JLabel(tempArray[i] + "\u00b0" + "F");
+				templabel[i] = new JLabel((int)tempArray[i] + "\u00b0" + "F");
 				templabel[i].setFont(minorfont);
 			} else {
-				templabel[i] = new JLabel(tempArray[i] + "\u00b0" + "C");
+				templabel[i] = new JLabel((int)tempArray[i] + "\u00b0" + "C");
 				templabel[i].setFont(minorfont);
 			}
